Move rpc.StateUpdate to its own file
